+++ src/parser/ldp/lomse_ldp_analyser.cpp
@@ -3925,13 +3925,8 @@ public:
         //tuplet
         if (fAddOldTuplet)
             set_old_tuplet(pNR);
-        else if (m_pTupletInfo==NULL
-                 && m_pAnalyser->is_tuplet_open()
-                 && m_pAnalyser->get_score_version() <= 107
-                )
-            add_to_current_tuplet(pNR);
-
-        add_tuplet_info(pNR);
+        else
+            add_tuplet_info(pNR);
 
         //time modification
         if (m_pTimeModifDto != NULL)
@@ -3969,22 +3964,6 @@ public:
       //  }
 
             ImoTreeAlgoritms::add_note_to_chord(pStartOfChordNote, pNote, pDoc);
-
-//            ImoChord* pChord;
-//            if (pStartOfChordNote->is_in_chord())
-//            {
-//                //chord already created. just add note to it
-//                pChord = pStartOfChordNote->get_chord();
-//            }
-//            else
-//            {
-//                //previous note is the base note. Create the chord
-//                pChord = static_cast<ImoChord*>(ImFactory::inject(k_imo_chord, pDoc));
-//                pStartOfChordNote->include_in_relation(pDoc, pChord);
-//            }
-//
-//            //add current note to chord
-//            pNote->include_in_relation(pDoc, pChord);
         }
 
         //save this note as last note
@@ -4159,14 +4138,6 @@ protected:
         }
     }
 
-    void add_to_current_tuplet(ImoNoteRest* pNR)
-    {
-        ImoTupletDto* pInfo = LOMSE_NEW ImoTupletDto();
-        pInfo->set_note_rest(pNR);
-        pInfo->set_tuplet_type(ImoTupletDto::k_continue);
-        m_pAnalyser->add_relation_info(pInfo);
-    }
-
     void set_old_tuplet(ImoNoteRest* pNR)
     {
         string value = m_srcOldTuplet;
@@ -4212,7 +4183,6 @@ protected:
             {
                 report_msg(m_pParamToAnalyse->get_line_number(),
                     "Requesting to start a tuplet but there is already an open tuplet. Tuplet ignored.");
-                add_to_current_tuplet(pNR);
             }
             else
             {
@@ -4249,6 +4219,7 @@ protected:
             //    pInfo->set_normal_number(0);  //required
         }
         ImoTupletDto* pInfo = LOMSE_NEW ImoTupletDto();
+        pInfo->set_tuplet_type(ImoTupletDto::k_start);
         pInfo->set_note_rest(pNR);
         pInfo->set_actual_number(actual);
         pInfo->set_normal_number(normal);
@@ -4285,9 +4256,17 @@ protected:
     {
         if (m_pTupletInfo)
         {
+            if (m_pTupletInfo->is_start_of_tuplet())
+                m_pAnalyser->add_to_open_tuplets(pNR);
+
             m_pTupletInfo->set_note_rest(pNR);
             m_pAnalyser->add_relation_info(m_pTupletInfo);
+
+            if (m_pTupletInfo->is_end_of_tuplet())
+                m_pAnalyser->add_to_open_tuplets(pNR);
         }
+        else
+            m_pAnalyser->add_to_open_tuplets(pNR);
     }
 
     void add_beam_info(ImoNoteRest* pNR)
@@ -6078,8 +6057,23 @@ public:
         set_default_values(pInfo);
 
         // [<tupletID>]     //optional. Only required for nested tuplets
-        if (get_optional(k_number))
-            set_tuplet_id(pInfo);
+        // <tupletID>   Mandatory since 2.1, for nested tuplets
+        // if (m_pAnalyser->get_score_version() > 200)
+        //{
+        //    if (get_mandatory(k_number))
+        //        set_tuplet_id(pInfo);
+        //    else
+        //    {
+        //        error_msg("Missing or invalid tuplet number. Tuplet ignored.");
+        //        delete pInfo;
+        //        return;
+        //    }
+        //}
+        //else
+        {
+            if (get_optional(k_number))
+                set_tuplet_id(pInfo);
+        }
 
         // { + | - }
         if (!get_mandatory(k_label) || !set_tuplet_type(pInfo))
@@ -7467,24 +7461,39 @@ void TupletsBuilder::add_relation_to_notes_rests(ImoTupletDto* pEndDto)
     ImoTupletDto* pStartDto = m_matches.front();
     m_pTuplet = ImFactory::inject_tuplet(pDoc, pStartDto);
 
-    //if not only graphical (LDP < 2.0) add time modification to all note/rest
-    if (!pStartDto->is_only_graphical())
+    //loop for including the tuplet in all note/rest
+    std::list<ImoTupletDto*>::iterator it;
+    for (it = m_matches.begin(); it != m_matches.end(); ++it)
     {
-        std::list<ImoTupletDto*>::iterator it;
-        for (it = m_matches.begin(); it != m_matches.end(); ++it)
-        {
-            ImoNoteRest* pNR = (*it)->get_note_rest();
+        //add tuplet to the note/rest
+        ImoNoteRest* pNR = (*it)->get_note_rest();
+        pNR->include_in_relation(pDoc, m_pTuplet, NULL);
+
+        //if not only graphical (LDP < 2.0) add time modification the note/rest
+        if (!pStartDto->is_only_graphical())
             pNR->set_time_modification( m_pTuplet->get_normal_number(),
                                         m_pTuplet->get_actual_number() );
+    }
+}
+
+//---------------------------------------------------------------------------------------
+void TupletsBuilder::add_to_open_tuplets(ImoNoteRest* pNR)
+{
+    if (m_pendingItems.size() > 0)
+    {
+        ListIterator it;
+        for(it=m_pendingItems.begin(); it != m_pendingItems.end(); ++it)
+        {
+            if ((*it)->is_start_of_relation() )
+            {
+                ImoTupletDto* pInfo = LOMSE_NEW ImoTupletDto();
+                pInfo->set_tuplet_number( (*it)->get_item_number() );
+                pInfo->set_tuplet_type(ImoTupletDto::k_continue);
+                pInfo->set_note_rest(pNR);
+                save_item_info(pInfo);
+            }
         }
     }
-
-    //create tuplet data and add tuplet to start and end note/rest
-    ImoNoteRest* pNR = pStartDto->get_note_rest();
-    pNR->include_in_relation(pDoc, m_pTuplet, NULL);
-
-    pNR = pEndDto->get_note_rest();
-    pNR->include_in_relation(pDoc, m_pTuplet, NULL);
 }